ignore db hits during AA load
JFC I HATE THIS WORKAROUND

I'm not sure that it's possible to have a clean solution for this though?
During migrations, I need to load Rails. AA is part of that load. AA will *always* fail because certain models reference other models before their tables exist.

Rawr. I hope AA 2.0 has a solution for this, and/or I think of something clever.